refactor: rename `Destinations` page to `ListDestinations`
