+++ web/src/pages/ListDestinations/Skeleton/Skeleton.tsx
