+++ web/src/pages/ListDestinations/Skeleton/index.tsx
