+++ web/src/pages/ListDestinations/graphql/listDestinationsAndDefaults.generated.ts
