+++ web/src/pages/ListDestinations/graphql/listDestinationsAndDefaults.graphql
