+++ web/src/routes.tsx
@@ -25,7 +25,7 @@ import ResourceDetailsPage from 'Pages/ResourceDetails';
 import PolicyDetailsPage from 'Pages/PolicyDetails';
 import GeneralSettingsPage from 'Pages/GeneralSettings';
 import SignInPage from 'Pages/SignIn';
-import DestinationsPage from 'Pages/Destinations';
+import ListDestinationsPage from 'Pages/ListDestinations';
 import UsersPage from 'Pages/Users';
 import RuleDetailsPage from 'Pages/RuleDetails';
 import LandingPage from 'Pages/Landing';
@@ -202,7 +202,7 @@ const PrimaryPageLayout: React.FunctionComponent = () => {
                   to={urls.settings.globalPythonModules.edit(':id')}
                 />
                 <Route exact path={urls.settings.users()} component={UsersPage} />
-                <Route exact path={urls.settings.destinations()} component={DestinationsPage} />
+                <Route exact path={urls.settings.destinations()} component={ListDestinationsPage} />
                 <Route component={Page404} />
               </Switch>
             </APIErrorFallback>
